+++ src/ConditionUserInRole.php
@@ -21,7 +21,7 @@ final readonly class ConditionUserInRole implements BookPageCondition
      * @param string $parameter Role
      * @throws TypeError
      */
-    public function isAllowed($parameter = null): bool
+    public function isAllowed(mixed $parameter = null): bool
     {
         if (!is_string($parameter)) {
             throw new TypeError("Method " . __METHOD__ . " expects string as parameter.");